release.sh: Compress with gzip
As this is a crucial mini-filesystem for InkBox OS to start up and it contains the essential binaries to mount other SquashFSes with FUSE instead of a native kernel driver, it needs to be compressed with gzip, otherwise some devices with a <3.x kernel won't support mounting it natively as it would have been compressed with xz, and thus the whole OS will crash.
This commit hopefully fixes this issue.